# internal/pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cartomancer API
  version: 0.1.0
  description: |
    REST API for managing GPX/FIT tracks for cycling and running.

    ## Authentication

    All requests pass through the session middleware, which auto-creates an
    anonymous session (JWT in an HTTP-only, SameSite=Strict cookie).
    Logging in promotes the session to an authenticated one.

    ### Access levels

    | Level      | Description                                              |
    |------------|----------------------------------------------------------|
    | **public** | No authentication required. Anonymous sessions suffice.  |
    | **user**   | Requires a logged-in user.                    |
    | **admin**  | Requires a logged-in user with `admin = true`.           |

    Endpoints that require **user** access return `401` if the session is
    anonymous. Endpoints that require **admin** access return `401` if
    anonymous, or `403` if the user is not an admin.
servers:
  - url: /api
    description: Same-origin
security:
  - cookieAuth: []
...
paths:
...
  /admin/segments/{uuid}:
    get:
      summary: Get segment details
      operationId: getSegment
      tags: [Segments]
      description: >
        Returns the detail of a single segment including its junctions,
        polyline, and member track UUIDs. Admin only.
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Segment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegmentDetailResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Segment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /tags:
    get:
      summary: Suggest tags by prefix
      operationId: suggestTags
      tags: [Tags]
      parameters:
        - name: prefix
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        "200":
          description: Tag suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagSuggestionResponse"
        "400":
          description: Prefix too short
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: >
        JWT session token set automatically by the session middleware.
        Logging in via POST /sessions/login promotes the anonymous session
        to an authenticated one. The cookie is HTTP-only, SameSite=Strict.
  schemas:
...
    TagSuggestionResponse:
      type: object
      required:
        - tags
      properties:
        tags:
          type: array
          items:
            type: string
    Error:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
...
    SegmentJunction:
      type: object
      required:
        - uuid
        - h3Cell
        - lat
        - lon
      properties:
        uuid:
          type: string
        h3Cell:
          type: string
        lat:
          type: number
        lon:
          type: number
    SegmentDetailResponse:
      type: object
      required:
        - uuid
        - distanceM
        - ascentM
        - nTracks
        - h3Resolution
        - polyline
        - startJunction
        - endJunction
        - trackUuids
      properties:
        uuid:
          type: string
        distanceM:
          type: number
        ascentM:
          type: number
        nTracks:
          type: integer
        h3Resolution:
          type: integer
        polyline:
          type: string
        startJunction:
          $ref: "#/components/schemas/SegmentJunction"
        endJunction:
          $ref: "#/components/schemas/SegmentJunction"
        trackUuids:
          type: array
          items:
            type: string
...